Move the Now effect from Spar to a library
- This effect is generally useful and it will be used in services other
than Spar
