+++ services/spar/package.yaml
@@ -56,6 +56,7 @@ dependencies:
   - polysemy
   - polysemy-check >= 0.9
   - polysemy-plugin
+  - polysemy-wire-zoo
   - QuickCheck
   - raw-strings-qq
   - retry
