+++ services/spar/src/Spar/Sem/AssIDStore/Cassandra.hs
@@ -32,9 +32,9 @@ import qualified SAML2.WebSSO as SAML
 import qualified Spar.Data as Data
 import Spar.Data.Instances ()
 import Spar.Sem.AssIDStore
-import Spar.Sem.Now (Now)
-import qualified Spar.Sem.Now as Now
 import Wire.API.User.Saml
+import Wire.Sem.Now (Now)
+import qualified Wire.Sem.Now as Now
 
 assIDStoreToCassandra ::
   forall m r a.
