+++ services/spar/src/Spar/Sem/BindCookieStore/Cassandra.hs
@@ -36,11 +36,11 @@ import SAML2.WebSSO (fromTime)
 import qualified SAML2.WebSSO as SAML
 import qualified Spar.Data as Data
 import Spar.Sem.BindCookieStore
-import Spar.Sem.Now (Now)
-import qualified Spar.Sem.Now as Now
 import qualified Web.Cookie as Cky
 import Wire.API.Cookie
 import Wire.API.User.Saml
+import Wire.Sem.Now (Now)
+import qualified Wire.Sem.Now as Now
 
 bindCookieStoreToCassandra ::
   forall m r a.
