+++ services/spar/src/Spar/Sem/VerdictFormatStore/Mem.hs
@@ -28,10 +28,10 @@ import Polysemy
 import Polysemy.State hiding (Get)
 import SAML2.WebSSO (addTime)
 import qualified SAML2.WebSSO.Types as SAML
-import Spar.Sem.Now (Now, boolTTL)
-import qualified Spar.Sem.Now as Now
 import Spar.Sem.VerdictFormatStore
 import Wire.API.User.Saml (AReqId, VerdictFormat)
+import Wire.Sem.Now (Now, boolTTL)
+import qualified Wire.Sem.Now as Now
 
 verdictFormatStoreToMem ::
   Member Now r =>
