+++ services/spar/test/Test/Spar/Sem/NowSpec.hs
@@ -28,11 +28,11 @@ import Imports
 import Polysemy
 import Polysemy.Input
 import SAML2.WebSSO.Types
-import Spar.Sem.Now.IO
-import Spar.Sem.Now.Input
-import Spar.Sem.Now.Spec
 import Test.Hspec
 import Test.Hspec.QuickCheck
+import Wire.Sem.Now.IO
+import Wire.Sem.Now.Input
+import Wire.Sem.Now.Spec
 
 someTime :: UTCTime
 someTime = UTCTime (fromJulianYearAndDay 1990 209) (secondsToDiffTime 0)
